effortles/requirements.txt
aiopg==1.3.3
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
black==22.3.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0
django-celery-beat==2.3.0
django-celery-results==2.3.1
django-debug-toolbar==3.4.0
django-filter==21.1
django-kronos==1.0
django-timezone-field==5.0
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
iniconfig==1.1.1
kombu==5.2.4
Markdown==3.3.7
model-bakery==1.5.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
redis==4.3.3
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
stripe==3.3.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
zipp==3.8.0
